+++ src/Bitcoin/SignatureBuilder.cpp
@@ -4,7 +4,7 @@
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
-#include "TransactionSigner.h"
+#include "SignatureBuilder.h"
 
 #include "SigHashType.h"
 #include "TransactionInput.h"
