+++ src/Bitcoin/SignatureBuilder.h
