+++ tests/DigiByte/TWDigiByteTests.cpp
@@ -8,7 +8,7 @@
 
 #include "Bitcoin/OutPoint.h"
 #include "Bitcoin/TransactionBuilder.h"
-#include "Bitcoin/TransactionSigner.h"
+#include "Bitcoin/SignatureBuilder.h"
 #include "HexCoding.h"
 #include "PublicKey.h"
